pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "seclab-taskflow-agent"
dynamic = ["version"]
description = "A taskflow agent for the SecLab project, enabling secure and automated workflow execution."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [
  { name = "GitHub Security Lab", email = "securitylab@github.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "aiofiles==24.1.0",
  "annotated-types==0.7.0",
  "anyio==4.9.0",
  "attrs==25.3.0",
  "Authlib==1.6.6",
  "certifi==2025.6.15",
  "cffi==2.0.0",
  "charset-normalizer==3.4.2",
  "click==8.2.1",
  "colorama==0.4.6",
  "cryptography==45.0.7",
  "cyclopts==3.24.0",
  "distro==1.9.0",
  "dnspython==2.8.0",
  "docstring_parser==0.17.0",
  "docutils==0.22",
  "email-validator==2.3.0",
  "exceptiongroup==1.3.0",
  "fastmcp==2.12.2",
  "griffe==1.7.3",
  "h11==0.16.0",
  "httpcore==1.0.9",
  "httpx==0.28.1",
  "httpx-sse==0.4.1",
  "idna==3.10",
  "importlib_metadata==8.7.0",
  "isodate==0.7.2",
  "jedi==0.19.2",
  "jiter==0.10.0",
  "jsonschema==4.24.0",
  "jsonschema-path==0.3.4",
  "jsonschema-specifications==2025.4.1",
  "lazy-object-proxy==1.12.0",
  "markdown-it-py==3.0.0",
  "MarkupSafe==3.0.2",
  "mcp==1.13.1",
  "mdurl==0.1.2",
  "more-itertools==10.8.0",
  "openai==1.107.0",
  "openai-agents==0.2.11",
  "openapi-core==0.19.5",
  "openapi-pydantic==0.5.1",
  "openapi-schema-validator==0.6.3",
  "openapi-spec-validator==0.7.2",
  "parse==1.20.2",
  "parso==0.8.4",
  "pathable==0.4.4",
  "platformdirs==4.5.0",
  "pluggy==1.6.0",
  "pycparser==2.23",
  "pydantic==2.11.7",
  "pydantic-settings==2.10.1",
  "pydantic_core==2.33.2",
  "Pygments==2.19.2",
  "pyperclip==1.9.0",
  "python-dotenv==1.1.1",
  "python-lsp-jsonrpc==1.1.2",
  "python-lsp-server==1.12.2",
  "python-multipart==0.0.22",
  "PyYAML==6.0.2",
  "referencing==0.36.2",
  "requests==2.32.4",
  "rfc3339-validator==0.1.4",
  "rich==14.0.0",
  "rich-rst==1.3.1",
  "rpds-py==0.26.0",
  "shellingham==1.5.4",
  "six==1.17.0",
  "sniffio==1.3.1",
  "SQLAlchemy==2.0.41",
  "sse-starlette==2.4.1",
  "starlette==0.49.1",
  "strenum==0.4.15",
  "tqdm==4.67.1",
  "typer==0.16.0",
  "types-requests==2.32.4.20250611",
  "typing-inspection==0.4.1",
  "typing_extensions==4.14.1",
  "ujson==5.10.0",
  "urllib3==2.6.3",
  "uvicorn==0.35.0",
  "zipp==3.23.0",
]

[project.urls]
Source = "https://github.com/GitHubSecurityLab/seclab-taskflow-agent"
Issues = "https://github.com/GitHubSecurityLab/seclab-taskflow-agent/issues"

[tool.hatch.version]
path = "src/seclab_taskflow_agent/__about__.py"

[tool.hatch.envs.default.scripts]
main = "python -m seclab_taskflow_agent {args:tests}"
sync-deps = ""

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/seclab_taskflow_agent tests}"

[tool.coverage.run]
source_pkgs = ["seclab_taskflow_agent", "tests"]
branch = true
parallel = true
omit = [
  "src/seclab_taskflow_agent/__about__.py",
]

[tool.coverage.paths]
seclab_taskflow_agent = ["src/seclab_taskflow_agent", "*/seclab-taskflow-agent/src/seclab_taskflow_agent"]
tests = ["tests", "*/seclab-taskflow-agent/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.ruff]
# Set target version to 3.10 to support match statements
target-version = "py310"

[tool.ruff.lint]
# Suppress all current linter errors to establish a baseline
ignore = [
  "A001",     # Variable shadows built-in
  "A002",     # Argument shadows built-in
  "A004",     # Import shadows built-in
  "ARG001",   # Unused function argument
  "B006",     # Mutable default argument
  "B007",     # Unused loop control variable
  "B008",     # Function call in argument defaults
  "B023",     # Function uses loop variable
  "B904",     # raise-without-from-inside-except
  "BLE001",   # Blind except
  "C405",     # Unnecessary literal set
  "C416",     # Unnecessary comprehension
  "E721",     # Type comparison using ==
  "E722",     # Bare except
  "E741",     # Ambiguous variable name
  "EM101",    # Exception string literal
  "EM102",    # Exception f-string
  "F541",     # f-string without placeholders
  "F811",     # Redefinition of unused name
  "F821",     # Undefined name
  "F841",     # Unused variable
  "FA100",    # Missing from __future__ import annotations
  "FA102",    # Missing from __future__ import annotations in stub
  "FBT001",   # Boolean positional arg in function definition
  "FBT002",   # Boolean default value in function definition
  "FURB188",  # Prefer removeprefix over conditional slice
  "G004",     # Logging with f-string
  "I001",     # Import block unsorted or unformatted
  "INP001",   # Implicit namespace package
  "LOG015",   # root logger usage
  "N801",     # Class name should use CapWords convention
  "N802",     # Function name should be lowercase
  "N806",     # Variable in function should be lowercase
  "N818",     # Exception name should end with Error
  "PERF102",  # Use keys() or values() instead of items()
  "PERF401",  # Use list comprehension
  "PIE790",   # Unnecessary pass statement
  "PLC0415",  # Import should be at top of file
  "PLC1802",  # Use of len(x) == 0
  "PLR2004",  # Magic value used in comparison
  "PLW0602",  # Global variable not assigned
  "PLW0603",  # Using global statement
  "PLW1508",  # Invalid envvar default
  "PLW2901",  # Outer loop variable overwritten
  "PT011",    # pytest.raises too broad
  "PYI041",   # Use float instead of int | float
  "RET503",   # Missing explicit return
  "RET504",   # Unnecessary assignment before return
  "RET505",   # Unnecessary else after return
  "RET506",   # Unnecessary else after raise
  "RUF005",   # Unpack instead of concatenation
  "RUF010",   # Use explicit conversion flag
  "RUF015",   # Prefer next() over single element slice
  "RUF059",   # Use of private function/attribute
  "RUF100",   # Unused noqa directive
  "S108",     # Hardcoded temp file/directory
  "S607",     # Starting process with partial path
  "SIM102",   # Use single if statement
  "SIM115",   # Use context handler for file
  "SIM210",   # Use ternary operator
  "SLF001",   # Private member access
  "T201",     # print found
  "TID252",   # Relative imports from parent modules
  "TRY003",   # Raise vanilla args
  "TRY004",   # Prefer TypeError for wrong type
  "TRY300",   # Consider moving statement to else
  "TRY301",   # Abstract raise to inner function
  "TRY400",   # Use logging.exception instead of logging.error
  "UP004",    # Use X | Y for union types
  "UP006",    # Use X | Y for union types in isinstance
  "UP009",    # UTF-8 encoding declaration
  "UP015",    # Unnecessary mode argument
  "UP020",    # Use builtin open
  "UP024",    # Replace aliased errors with OSError
  "UP032",    # Use f-string
  "UP035",    # Import from collections.abc
  "UP045",    # Use X | None for type annotations
  "W291",     # Trailing whitespace
  "W293",     # Blank line contains whitespace
]
